Fix(1). 77850 Team avatar upload. Team review.

+++ src/main/java/faang/school/projectservice/service/TeamAvatarServiceImpl.java
@@ -3,18 +3,13 @@ package faang.school.projectservice.service;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.List;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import faang.school.projectservice.config.context.UserContext;
 import faang.school.projectservice.exception.AuthorizationException;
 import faang.school.projectservice.model.Team;
-import faang.school.projectservice.model.TeamMember;
-import faang.school.projectservice.model.TeamRole;
-import faang.school.projectservice.repository.TeamMemberRepository;
 import faang.school.projectservice.repository.TeamRepository;
 import faang.school.projectservice.service.s3.S3Service;
 import jakarta.persistence.EntityNotFoundException;
@@ -25,11 +20,10 @@ import net.coobird.thumbnailator.Thumbnails;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class TeamAvatarUploadServiceImpl {
+public class TeamAvatarServiceImpl implements TeamAvatarService {
     private final TeamRepository teamRepository;
-    private final TeamMemberRepository teamMemberRepository;
     private final S3Service s3Service;
-    private final UserContext userContext;
+    private final TeamMemberService teamMemberService;
 
     @Value("${team-avatar-file.maxDimension}")
     private int avatarMaxDimension;
@@ -48,14 +42,15 @@ public class TeamAvatarUploadServiceImpl {
         return s3Service.downloadFile(team.getAvatarKey());
     }
 
-    public void deleteFile(long teamId) {
-        List<TeamMember> teamMembers = teamMemberRepository.findByUserId(userContext.getUserId());        
-        teamMembers.stream()
-            .filter(member -> member.getTeam().getId().equals(teamId))
-            .filter(member -> member.getRoles().contains(TeamRole.MANAGER))
-            .findFirst()
-            .orElseThrow(() -> new AuthorizationException(String.format("You are not a manager of this team %d.", teamId)));
+    public String getAvatarContentType(long teamId) {
+        Team team = getTeam(teamId);
+        return s3Service.getAvatarContentType(team.getAvatarKey());
+    }
 
+    public void deleteFile(long teamId) {
+        if (!teamMemberService.ifUserIsManager(teamId)) {
+            throw new AuthorizationException(String.format("You are not a manager of this team %d.", teamId));
+        }
         Team team = getTeam(teamId);
         s3Service.deleteFile(team.getAvatarKey());
         team.setAvatarKey(null);
